examples/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
anthropic>=0.50
openai>=1.30
